Moved time consuming test case to own test class that has been annoted with "ignore".
